♻️ Refactor(#12): 프로토콜 파일 분리
