+++ secondhand/secondhand/Protocol/UserProtocols/UserSignUpProvider.swift
@@ -1,0 +1,13 @@
+//
+//  UserSignUpProvider.swift
+//  secondhand
+//
+//  Created by 조호근 on 6/25/24.
+//
+
+import Foundation
+
+protocol UserSignUpProvider {
+    func isUserNameTaken(_ username: String) -> Bool
+    func saveUser(_ user: User) throws
+}
